# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Torneos",
        "description": "API para gestionar torneos de tenis",
        "version": "1.0.0"
    },
    "paths": {
        "/api/simular-torneo": {
            "post": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Simula y guarda un torneo",
                "operationId": "0e4f64863b6d832968ff2592a9cac4d6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jugadores": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Jugador"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Torneo simulado y guardado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Torneo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "/api/buscar-torneos-por-fecha": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Busca torneos por fecha",
                "operationId": "b552d343442629a8c53ccf0578ebd659",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de torneos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Torneo"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Formato de fecha inválido"
                    }
                }
            }
        },
        "/api/buscar-torneos-por-genero": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Busca torneos por género",
                "operationId": "98f3ff73c7adda22532220ffd7d94db2",
                "parameters": [
                    {
                        "name": "genero",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de torneos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Torneo"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "/api/obtener-todos-los-torneos": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Obtiene todos los torneos",
                "operationId": "b46eed37b1fd54aded469004baa0d19c",
                "responses": {
                    "200": {
                        "description": "Lista de todos los torneos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Torneo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/buscar-torneo-por-id": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Busca un torneo por ID",
                "operationId": "a03b0cea1b3e5d2777c19246e384f202",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Torneo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Torneo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Torneo no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Jugador": {
                "required": [
                    "nombre",
                    "edad",
                    "genero",
                    "nivel_habilidad"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del jugador",
                        "type": "string"
                    },
                    "edad": {
                        "description": "Edad del jugador",
                        "type": "integer"
                    },
                    "genero": {
                        "description": "Género del jugador",
                        "type": "string",
                        "enum": [
                            "masculino",
                            "femenino"
                        ]
                    },
                    "nivel_habilidad": {
                        "description": "Nivel de habilidad del jugador",
                        "type": "integer"
                    },
                    "tiempo_reaccion": {
                        "description": "Tiempo de reacción (solo para jugadoras femeninas)",
                        "type": "number",
                        "format": "float"
                    },
                    "fuerza": {
                        "description": "Fuerza del jugador (solo para jugadores masculinos)",
                        "type": "integer"
                    },
                    "velocidad": {
                        "description": "Velocidad del jugador (solo para jugadores masculinos)",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Torneo": {
                "properties": {
                    "id": {
                        "description": "ID del torneo",
                        "type": "integer",
                        "format": "int64"
                    },
                    "fecha": {
                        "description": "Fecha del torneo",
                        "type": "string",
                        "format": "date"
                    },
                    "genero": {
                        "description": "Género del torneo",
                        "type": "string",
                        "enum": [
                            "masculino",
                            "femenino",
                            "mixto"
                        ]
                    },
                    "ganador": {
                        "$ref": "#/components/schemas/Jugador"
                    },
                    "participantes": {
                        "description": "Lista de jugadores participantes",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Jugador"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Torneos",
            "description": "API Endpoints de torneos"
        }
    ]
}